provide columnDef to cellTemplate
Otherwise, cellTemplate need access columnDef's info out of the call back function's closure. The confusion is that the variable that a function accesses is evaluated when the function is invoked rather than when it is defined. This can be resolve by defining an immediately invoked function expression (IIFE). But, provide columnDef as a parameter is just easier.